refactor: remove old account conversion methods
